Renaming lots of files due to STI implementation of profile
