+++ app/controllers/financial_profiles_controller.rb
@@ -1,0 +1,18 @@
+class FinancialProfilesController < ProfilesController
+  def new
+    @profile = FinancialProfile.new
+  end
+
+  #######
+  private
+  #######
+
+  def profile_scope
+    :financial_profile
+  end
+
+  def fetch_profile
+    @profile = FinancialProfile.find(params[:id])
+    authorize @profile
+  end
+end
