+++ app/models/financial_profile.rb
@@ -1,0 +1,3 @@
+class FinancialProfile < Profile
+  belongs_to :user
+end
