+++ app/models/financial_profile_policy.rb
@@ -1,0 +1,3 @@
+class FinancialProfilePolicy < ApplicationPolicy
+  include PrivatePolicy
+end
