+++ app/models/user.rb
@@ -3,7 +3,7 @@ class User < ActiveRecord::Base
 
   has_many :testing_grounds
   has_many :topologies
-  has_many :load_profiles
+  has_many :profiles
   has_many :market_models
 
   def activate!
