+++ app/services/technology_profiles/query.rb
@@ -1,0 +1,24 @@
+module TechnologyProfiles
+  class Query
+    def initialize(distribution)
+      @distribution = distribution
+    end
+
+    def query
+      Hash[permits.group_by(&:technology).map do |tech_key, techs|
+        [tech_key, techs.map { |tech| tech.load_profile.key }]
+      end]
+    end
+
+    private
+
+      def technology_keys
+        @distribution.map{|t| t['type']}.uniq
+      end
+
+      def permits
+        TechnologyProfile.where(technology: technology_keys)
+                          .includes(:load_profile)
+      end
+  end
+end
