+++ app/views/financial_profiles/_form.html.haml
@@ -1,0 +1,43 @@
+- if @profile.errors.any?
+  %ul.errors
+    - @profile.errors.full_messages.uniq.each do |message|
+      %li= message
+
+= form_for @profile, html: { multipart: true } do |form|
+  - if @profile.new_record?
+    .form-group
+      = form.label :key
+      = tooltip_tag(t(:'load_profiles.form.key.tooltip'))
+      = form.text_field :key, class: 'form-control'
+      .help-block This field is optional.
+
+  .form-group
+    = form.label :name
+    = form.text_field :name, class: 'form-control'
+    .help-block This field is optional.
+
+  .form-group
+    = form.label :public, "Permissions"
+
+    .radio-buttons
+      = form.radio_button :public, true
+      = form.label :public, "Public", class: "radio-label"
+      = form.radio_button :public, false
+      = form.label :public, "Private", class: "radio-label"
+
+  - unless @profile.locked?
+    .form-group
+      = form.label :curve
+      = tooltip_tag(t(:'load_profiles.form.curve.tooltip'))
+      = form.file_field :curve, accept: 'text/csv'
+
+      - unless @profile.new_record?
+        .help-block Leave blank if you don't want to change the curve.
+
+  .form-group
+    = form.submit class: 'btn btn-primary'
+    %span.or or
+    - if @profile.new_record?
+      = link_to 'Cancel', financial_profiles_path, class: 'btn btn-default'
+    - else
+      = link_to 'Cancel', financial_profile_path(@profile), class: 'btn btn-default'
