+++ app/views/financial_profiles/new.html.haml
@@ -1,0 +1,2 @@
+%h3== Create a new financial profile
+= render partial: 'form'
