+++ app/views/layouts/_nav.html.haml
@@ -30,11 +30,12 @@
 
         %li.dropdown
           %a.dropdown-toggle{"aria-expanded" => "false", "data-toggle" => "dropdown", :href => "#", :role => "button"}
-            Load Profiles
+            Profiles
             %span.caret
           %ul.dropdown-menu{:role => "menu"}
             %li
-              = link_to 'Create new', new_load_profile_path
+              = link_to 'Create new load profile', new_load_profile_path
+              = link_to 'Create new financial profile', new_financial_profile_path
             %li.divider
             %li
               = link_to 'Show all', load_profiles_path
