+++ app/views/load_profiles/index.html.haml
@@ -1,4 +1,8 @@
-%h1 List of all Load Profiles available
+%h1 List of all profiles available
+
+%h2 Load Profiles
 
 .load_profiles
   = render partial: "load_profile_categories/load_profile_categories", locals: { load_profile_categories: @load_profile_categories, collapsed: false }
+
+%h2 Financial Profiles
