+++ app/views/load_profiles/show.html.haml
@@ -1,14 +1,14 @@
 %ul
-  %li== Key: #{ @load_profile.key }
-  %li== Curve file name: #{ @load_profile.curve_file_name }
-  %li== Permitted Technologies: #{ @load_profile.technology_profiles.map(&:technology).join(", ") }
+  %li== Key: #{ @profile.key }
+  %li== Curve file name: #{ @profile.curve_file_name }
+  %li== Permitted Technologies: #{ @profile.technology_profiles.map(&:technology).join(", ") }
 
-.profile-graph.line-chart{ 'data-url' => load_profile_path(@load_profile, format: :json) }
+.profile-graph.line-chart{ 'data-url' => load_profile_path(@profile, format: :json) }
 
 .commit
-  = link_to('Download', @load_profile.curve.url, class: 'button')
+  = link_to('Download', @profile.curve.url, class: 'button')
   %span.or or
-  = link_to_if(policy(@load_profile).edit?, 'Edit', edit_load_profile_path(@load_profile), class: 'button')
-  = link_to_if(policy(@load_profile).destroy?, 'Delete', load_profile_path(@load_profile), class: 'button', method: 'delete', data: { confirm: 'Are you sure you want to delete this profile?' })
+  = link_to_if(policy(@profile).edit?, 'Edit', edit_load_profile_path(@profile), class: 'button')
+  = link_to_if(policy(@profile).destroy?, 'Delete', load_profile_path(@profile), class: 'button', method: 'delete', data: { confirm: 'Are you sure you want to delete this profile?' })
 
 = link_to '◄ All load profiles', load_profiles_path, class: 'button button-primary'
