+++ db/migrate/20150630090016_rename_load_profiles_to_profiles.rb
@@ -1,0 +1,5 @@
+class RenameLoadProfilesToProfiles < ActiveRecord::Migration
+  def change
+    rename_table :load_profiles, :profiles
+  end
+end
