+++ db/migrate/20150630090104_add_type_column_to_profiles.rb
@@ -1,0 +1,5 @@
+class AddTypeColumnToProfiles < ActiveRecord::Migration
+  def change
+    add_column :profiles, :type, :string, after: :id
+  end
+end
