+++ db/migrate/20150630090208_change_profile_types_to_load_profile.rb
@@ -1,0 +1,5 @@
+class ChangeProfileTypesToLoadProfile < ActiveRecord::Migration
+  def change
+    Profile.update_all(type: "LoadProfile")
+  end
+end
