+++ spec/controllers/financial_profiles_controller_spec.rb
@@ -1,0 +1,13 @@
+require 'rails_helper'
+
+RSpec.describe FinancialProfilesController do
+  it '#new path' do
+    get :new
+  end
+
+  it '#create a new financial profile' do
+  end
+
+  it '#updates an existing financial profile' do
+  end
+end
